requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==2.1.0
pandas==2.1.3
openpyxl==3.1.2
fpdf==1.7.2
jinja2==3.1.2